Documentation was added to every class.
Standard puppet documentation was added to each class. Of note is
the universe.pp file which included documentation from the galaxy
project's universe_wsgi.ini file.

Authorship as noted in the puppet class tries to accurately reflect who
worked on which files, so if issues are raised it might be clear which
developer to inform.