Large rewrite to use hiera and first_run until common_startup works
Paths were hardcoded in common_startup, so temporarily reverting to
first_run style configuration. Additionally, swapped out most data for
hiear config. If this is not how you all were intending to use hiera,
please revert it! I just wanted to get a VM actually provisioned with
puppet-galaxy, and hiera had to be included somewhere. For convenience
I did it in the galaxy module, but if it should be in g-r-p module, we
can revert this and move it there.